requirements.txt
fastapi==0.115.0
pillow==10.4.0
pyinstaller==6.2.0
pynput==1.7.7
pywebview==5.2
sqlmodel==0.0.22
uvicorn==0.31.0
websockets==14.1
